Test case for concurrency problem
App create instance of `RequestVote.Request` that is `ConcurrentValue`.
Then it creates a bunch of tasks with `Task.withGroup` and use instance
inside the block. Inside the block struct keep right values for
properties, but when it pass into a async method properties corrupted.

To reproduce it open the project and run `local-cluster` target. After
a moment app will break on assert.